Fixed refactoring conflict bug for UBUD rundown button
