+++ client/src/shared/components/grid/GridItem.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Box, Grid } from '@mui/material';
 import Button from 'shared/components/button/Button';
+import ReactGA from 'react-ga4';
 import GridItemDescription from 'shared/components/grid/GridItemDescription';
 import { useMediaQuery } from 'react-responsive';
 
@@ -9,6 +10,13 @@ export default function GridItem({ item, upcomingEvent, itemType }) {
 		window.open(`https://${link}`, '_blank', 'noreferrer');
 	};
 	const isMobileView = useMediaQuery({ query: '(max-width: 750px)' });
+	const handleRegisterButton = (link) => {
+		ReactGA.event({
+			category: 'Event',
+			action: 'Clicked register for UBUD',
+		});
+		window.open(`https://${link}`, '_blank', 'noreferrer');
+	};
 
 	return itemType === 'rantangan' && upcomingEvent ? (
 		<Grid item>
@@ -26,14 +34,18 @@ export default function GridItem({ item, upcomingEvent, itemType }) {
 	) : (
 		<Grid item xs={1} sm={1.5} md={2}>
 			<Box className='flex flex-col mx-7 sm:mx-0'>
-				<img src={item.image} alt='item_image' className='rounded-2xl' style={{ zIndex: 5 }} />
+				<img src={item.image} alt='item_image' className='rounded-2xl' />
 				<Box className='text-center flex flex-col py-4'>
 					<span className='py-2 font-semibold text-xl'>{item.title}</span>
-					<GridItemDescription item={item} itemType={itemType} />
+					<span>{item.date}</span>
+					<span>{item.loc}</span>
 				</Box>
 				{upcomingEvent && (
-					<Box className='flex justify-center'>
-						<Button text='Register' handleClickButton={() => handleClickButton(item.link)} />
+					<Box className='flex flex-col items-center justify-center space-y-6'>
+						<Button text='Register' handleClickButton={() => handleRegisterButton(item.registerLink)} />
+						<a href={item.infoLink} target='_blank' rel='noreferrer'>
+							<Button text='View Event Details' background='transparentBg' />
+						</a>
 					</Box>
 				)}
 			</Box>
